wip - trying to fix CF keys

+++ cloud-atlas/Controllers/AtlasController.cs
@@ -1,6 +1,5 @@
 using cloud_atlas;
 using cloud_atlas.Entities.Models;
-using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -24,7 +23,6 @@ public class AtlasController : BaseController
         .ToListAsync();
 
         return Ok(atlases);
-
     }
 
     [HttpPost]
@@ -59,9 +57,12 @@ public class AtlasController : BaseController
         var markerPhotosToRemove = await cosmosDbContext.MarkerPhotos
             .Where(mp => mp.AtlasId == dto.AtlasId).ToListAsync();
 
-        cosmosDbContext.MarkerPhotos.RemoveRange(markerPhotosToRemove);
+        if (markerPhotosToRemove.Any())
+        {
+            cosmosDbContext.MarkerPhotos.RemoveRange(markerPhotosToRemove);
 
-        await cosmosDbContext.SaveChangesAsync();
+            await cosmosDbContext.SaveChangesAsync();
+        }
 
         return Ok();
     }
@@ -69,7 +70,7 @@ public class AtlasController : BaseController
     [HttpPut]
     public async Task<IActionResult> UpdateAtlas([FromBody] UpdateAtlasDto updateAtlas)
     {
-        var IsOwner = await IsUserOwnerOfAtlas(updateAtlas.Id);
+        var IsOwner = await sqlDbContext.UserOwnsMap(updateAtlas.Id);
 
         if (!IsOwner) return Unauthorized();
 
@@ -90,33 +91,13 @@ public class AtlasController : BaseController
     [HttpPost("share")]
     public async Task<IActionResult> AddUserToAtlas([FromBody] AddUserToAtlasDto dto)
     {
-        var IsOwner = await IsUserOwnerOfAtlas(dto.AtlasId);
+        var IsOwner = await sqlDbContext.UserOwnsMap(dto.AtlasId);
 
         if (!IsOwner) return Unauthorized();
-
-        var atlasUser = await sqlDbContext
-        .AtlasUsers
-        .Include(au => au.Atlas)
-        .FirstOrDefaultAsync(au => au.AtlasId == dto.AtlasId && au.UserId == new Guid(HttpContext.Items["sub"] as string));
-
-        if (atlasUser is null) return NotFound();
-
-        if (atlasUser.IsOwner == false) return Unauthorized();
 
         sqlDbContext.AtlasUsers.Add(new AtlasUser() { AtlasId = dto.AtlasId, UserId = dto.UserId, IsOwner = false });
         await sqlDbContext.SaveChangesAsync();
 
         return Ok();
-    }
-
-    private async Task<bool> IsUserOwnerOfAtlas(Guid atlasId)
-    {
-        var atlasUser = await sqlDbContext
-        .AtlasUsers
-        .Include(au => au.Atlas)
-        .FirstOrDefaultAsync(au => au.AtlasId == atlasId && au.UserId == new Guid(HttpContext.Items["sub"] as string));
-
-        if (atlasUser.IsOwner == false) return false;
-        return true;
     }
 }
